requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.2
openpyxl==3.1.5
opt_einsum==3.4.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-irt==0.6.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyirt==0.3.4
pymongo==4.10.1
pyparsing==3.2.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.1
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
websockets==14.1
